Extract internal single-use trait
