+++ packages/framework/src/Framework/Features/Publications/Models/PublicationFields/PublicationFieldValue.php
@@ -7,9 +7,9 @@ declare(strict_types=1);
 namespace Hyde\Framework\Features\Publications\Models\PublicationFields;
 
 use DateTime;
+use Hyde\Framework\Features\Publications\Concerns\ParsesPublicationFieldInputs;
 use Hyde\Framework\Features\Publications\PublicationFieldTypes;
 use Hyde\Framework\Features\Publications\Validation\BooleanRule;
-use InvalidArgumentException;
 use function is_array;
 
 /**
@@ -21,6 +21,8 @@ use function is_array;
  */
 final class PublicationFieldValue
 {
+    use ParsesPublicationFieldInputs;
+
     public readonly PublicationFieldTypes $type;
     protected string|array|bool|float|int|DateTime $value;
 
@@ -71,85 +73,5 @@ final class PublicationFieldValue
             PublicationFieldTypes::Url => ['url'],
             PublicationFieldTypes::Tag => [],
         };
-    }
-
-    protected static function parseStringValue(string $value): string
-    {
-        return $value;
-    }
-
-    protected static function parseDatetimeValue(string $value): DateTime
-    {
-        return new DateTime($value);
-    }
-
-    protected static function parseBooleanValue(string $value): bool
-    {
-        return match ($value) {
-            'true', '1' => true,
-            'false', '0' => false,
-            default => throw self::parseError('boolean', $value)
-        };
-    }
-
-    protected static function parseIntegerValue(string $value): int
-    {
-        if (! is_numeric($value)) {
-            throw self::parseError('integer', $value);
-        }
-
-        return (int) $value;
-    }
-
-    protected static function parseFloatValue(string $value): float
-    {
-        if (! is_numeric($value)) {
-            throw self::parseError('float', $value);
-        }
-
-        return (float) $value;
-    }
-
-    protected static function parseImageValue(string $value): string
-    {
-        // TODO Validate file exists as the dynamic validation rules does the same
-        return $value;
-    }
-
-    protected static function parseArrayValue(string|array $value): array
-    {
-        return (array) $value;
-    }
-
-    protected static function parseTextValue(string $value): string
-    {
-        // In order to properly store multi-line text fields as block literals,
-        // we need to make sure the string ends with a newline character.
-        if (substr_count($value, "\n") > 0) {
-            return trim($value, "\r\n")."\n";
-        }
-
-        return $value;
-    }
-
-    protected static function parseUrlValue(string $value): string
-    {
-        if (! filter_var($value, FILTER_VALIDATE_URL)) {
-            throw self::parseError('url', $value);
-        }
-
-        return $value;
-    }
-
-    protected static function parseTagValue(string|array $value): array
-    {
-        return (array) $value;
-    }
-
-    protected static function parseError(string $typeName, string $input): InvalidArgumentException
-    {
-        return new InvalidArgumentException(sprintf("%s: Unable to parse invalid %s value '%s'",
-            (ucfirst($typeName).'Field'), $typeName, $input
-        ));
     }
 }
